projects/meridian/api/requirements.txt
fastapi
uvicorn
pydantic
pypdf
python-multipart
reportlab
requests
scikit-learn
openai
python-dotenv
